+++ src/obuf.c
@@ -88,7 +88,7 @@ void buf_escape(const char *str)
 
 void buf_add_str(const char *str)
 {
-    long i = 0;
+    size_t i = 0;
     while (str[i]) {
         if (!buf_put_char(u_str_get_char(str, &i))) {
             break;
@@ -167,7 +167,7 @@ static void skipped_too_much(unsigned int u)
         obuf.buf[obuf.count++] = '?';
     } else if (u_is_unprintable(u)) {
         char tmp[4];
-        long idx = 0;
+        size_t idx = 0;
         u_set_hex(tmp, &idx, u);
         memcpy(obuf.buf + obuf.count, tmp + 4 - n, n);
         obuf.count += n;
@@ -258,7 +258,7 @@ bool buf_put_char(unsigned int u)
             // <xx> would not fit.
             // There's enough space in the buffer so render all 4 characters
             // but increment position less.
-            long idx = obuf.count;
+            size_t idx = obuf.count;
             u_set_hex(obuf.buf, &idx, u);
             obuf.count += space;
             obuf.x += space;
